Fix bug dispatching fetchLogError and fetchLogSuccess
It turns out that I was not calling dispatch in my thunk pattern
in my actions/log where I was fetching and handling data for a part-
icular day. I ended up not using the Promise "catch" method and instead
checked to see if the snapshot existed using "snapshot.exist()" and if
it did, I dispatched success otherwise, I dispatch fail.